backend/requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==3.7.1
asgi-csrf==0.9
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
black==23.9.1
blinker==1.6.3
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Deprecated==1.2.14
dnspython==2.4.2
docstring-parser==0.15
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.99.1
fastapi-csrf-protect==0.3.2
fastapi-limiter==0.1.5
fastapi-users==12.1.2
Flask==3.0.0
Flask-Limiter==3.5.0
flask-secure-headers==0.6
gcloud==0.18.3
googleapis-common-protos==1.61.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
idna==3.4
importlib-resources==6.1.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
limits==3.6.0
makefun==1.15.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.1
oauth2client==4.1.3
ordered-set==4.1.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
piccolo==0.121.0
piccolo-api==0.58.0
platformdirs==3.11.0
pluggy==1.3.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
pytest==7.4.2
python-dotenv==1.0.0
python-jwt==4.0.0
python-multipart==0.0.6
pytz==2023.3.post1
redis==4.6.0
requests==2.29.0
requests-cache==1.1.0
requests-toolbelt==0.10.1
rich==13.6.0
rsa==4.9
sentry-sdk==1.32.0
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
starlette==0.27.0
stripe==6.7.0
targ==0.3.8
tomli==2.0.1
typing_extensions==4.8.0
url-normalize==1.4.3
urllib3==1.26.17
uvicorn==0.23.2
Werkzeug==3.0.0
wrapt==1.15.0
